+++ test/GCClassic/parallel/parallelTest.sh
@@ -5,14 +5,14 @@
 #------------------------------------------------------------------------------
 #BOP
 #
-# !MODULE: parTest.sh
+# !MODULE: paralleltest.sh
 #
 # !DESCRIPTION: Runs parallelization tests on the various GEOS-Chem Classic
 #  run directories (interactively, or with a scheduler).
 #\\
 #\\
 # !CALLING SEQUENCE:
-#  ./parTest.sh -d root-directory -e env-file [-l] [-p partition] [-q] [-s]
+#  ./parallelTest.sh -d root-directory -e env-file [-l] [-p partition] [-q] [-s]
 #
 #  Where the command-line arguments are as follows:
 #
@@ -53,8 +53,8 @@ quick="no"
 
 # Call Linux getopt function to specify short & long input options
 # (e.g. -d or --directory, etc).  Exit if not succesful
-validArgs=$(getopt --options d:e:hlp:qs \
-  --long directory:,env-file:,help,lsf,partition:,quick,slurm -- "$@")
+validArgs=$(getopt --options d:e:hp:qs: \
+  --long directory:,env-file:,help,partition:,quick,scheduler: -- "$@")
 if [[ $? -ne 0 ]]; then
     exit 1;
 fi
@@ -82,28 +82,22 @@ while [ : ]; do
             exit 1
             ;;
 
-	# -l or --lsf selects the LSF scheduler
-	-l | --lsf)
-	    scheduler="LSF"
-            shift
-            ;;
-
 	# -q or --quick runs a quick set of parallelization tests (for testing)
 	-q | --quick)
 	    quick="yes"
             shift
 	    ;;
 
-	# -p or --partition edits the SLURM and LSF tags
+	# -p or --partition replaces REQUESTED_PARTITON w/ the user's choice
 	-p | --partition)
 	    sedCmd="s/REQUESTED_PARTITION/${2}/"
 	    shift 2
 	    ;;
 
-	# -s or --slurm selects the SLURM scheduler
-	-s | --slurm)
-	    scheduler="SLURM"
-            shift
+	# -s or --scheduler selects the scheduler to use
+	-s | --scheduler)
+	    scheduler="${2^^}"
+            shift 2
             ;;
 
 	--) shift;
@@ -158,7 +152,7 @@ thisDir=$(pwd -P)
 ptRoot=$(absolute_path "${ptRoot}")
 
 # Create GEOS-Chem run directories in the parallelization test root folder
-./parTestCreate.sh "${ptRoot}" "${envFile}" "${quick}"
+./parallelTestCreate.sh "${ptRoot}" "${envFile}" "${quick}"
 if [[ $? -ne 0 ]]; then
    echo "ERROR: Could not create parallelization test run directories!"
    exit 1
@@ -182,38 +176,38 @@ if [[ "x${scheduler}" == "xSLURM" ]]; then
     #-------------------------------------------------------------------------
 
     # Remove LSF #BSUB tags
-    sed_ie '/#BSUB -q REQUESTED_PARTITION/d' "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#BSUB -n 8/d'                   "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#BSUB -W 00:30/d'               "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#BSUB -o lsf-%J.txt/d'          "${ptRoot}/parTestCompile.sh"
+    sed_ie '/#BSUB -q REQUESTED_PARTITION/d' "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#BSUB -n 8/d'                   "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#BSUB -W 00:30/d'               "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#BSUB -o lsf-%J.txt/d'          "${ptRoot}/parallelTestCompile.sh"
     sed_ie \
 	'/#BSUB -R "rusage\[mem=8GB\] span\[ptile=1\] select\[mem < 1TB\]"/d' \
-	"${ptRoot}/parTestCompile.sh"
+	"${ptRoot}/parallelTestCompile.sh"
     sed_ie \
 	"/#BSUB -a 'docker(registry\.gsc\.wustl\.edu\/sleong\/esm\:intel\-2021\.1\.2)'/d" \
-	"${ptRoot}/parTestCompile.sh"
-    sed_ie '/#BSUB -q REQUESTED_PARTITION/d' "${ptRoot}/parTestExecute.sh"
-    sed_ie '/#BSUB -n 24/d'                  "${ptRoot}/parTestExecute.sh"
-    sed_ie '/#BSUB -W 3:30/d'                "${ptRoot}/parTestExecute.sh"
-    sed_ie '/#BSUB -o lsf-%J.txt/d'          "${ptRoot}/parTestExecute.sh"
+	"${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#BSUB -q REQUESTED_PARTITION/d' "${ptRoot}/parallelTestExecute.sh"
+    sed_ie '/#BSUB -n 24/d'                  "${ptRoot}/parallelTestExecute.sh"
+    sed_ie '/#BSUB -W 3:30/d'                "${ptRoot}/parallelTestExecute.sh"
+    sed_ie '/#BSUB -o lsf-%J.txt/d'          "${ptRoot}/parallelTestExecute.sh"
     sed_ie \
 	'/#BSUB -R "rusage\[mem=90GB\] span\[ptile=1\] select\[mem < 2TB\]"/d' \
-	"${ptRoot}/parTestExecute.sh"
+	"${ptRoot}/parallelTestExecute.sh"
     sed_ie \
 	"/#BSUB -a 'docker(registry\.gsc\.wustl\.edu\/sleong\/esm\:intel\-2021\.1\.2)'/d" \
-	"${ptRoot}/parTestExecute.sh"
+	"${ptRoot}/parallelTestExecute.sh"
 
     # Replace "REQUESTED_PARTITION" with the partition name
-    sed_ie "${sedCmd}" "${ptRoot}/parTestCompile.sh"
-    sed_ie "${sedCmd}" "${ptRoot}/parTestExecute.sh"
+    sed_ie "${sedCmd}" "${ptRoot}/parallelTestCompile.sh"
+    sed_ie "${sedCmd}" "${ptRoot}/parallelTestExecute.sh"
 
     # Submit compilation tests script
-    output=$(sbatch parTestCompile.sh)
+    output=$(sbatch parallelTestCompile.sh)
     output=($output)
     cmpId=${output[3]}
 
     # Submit execution tests script as a job dependency
-    output=$(sbatch --dependency=afterok:${cmpId} parTestExecute.sh)
+    output=$(sbatch --dependency=afterok:${cmpId} parallelTestExecute.sh)
     output=($output)
     exeId=${output[3]}
 
@@ -228,33 +222,33 @@ elif [[ "x${scheduler}" == "xLSF" ]]; then
     #-------------------------------------------------------------------------
 
     # Remove SLURM #SBATCH tags
-    sed_ie '/#SBATCH -c 8/d'                   "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#SBATCH -N 1/d'                   "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#SBATCH -t 0-00:30/d'             "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#SBATCH -p REQUESTED_PARTITION/d' "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#SBATCH --mem=8000/d'             "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#SBATCH -p REQUESTED_PARTITION/d' "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#SBATCH --mail-type=END/d'        "${ptRoot}/parTestCompile.sh"
-    sed_ie '/#SBATCH -c 24/d'                  "${ptRoot}/parTestExecute.sh"
-    sed_ie '/#SBATCH -N 1/d'                   "${ptRoot}/parTestExecute.sh"
-    sed_ie '/#SBATCH -t 0-03:30/d'             "${ptRoot}/parTestExecute.sh"
-    sed_ie '/#SBATCH -p REQUESTED_PARTITION/d' "${ptRoot}/parTestExecute.sh"
-    sed_ie '/#SBATCH --mem=90000/d'            "${ptRoot}/parTestExecute.sh"
-    sed_ie '/#SBATCH --mail-type=END/d'        "${ptRoot}/parTestExecute.sh"
+    sed_ie '/#SBATCH -c 8/d'                   "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#SBATCH -N 1/d'                   "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#SBATCH -t 0-00:30/d'             "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#SBATCH -p REQUESTED_PARTITION/d' "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#SBATCH --mem=8000/d'             "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#SBATCH -p REQUESTED_PARTITION/d' "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#SBATCH --mail-type=END/d'        "${ptRoot}/parallelTestCompile.sh"
+    sed_ie '/#SBATCH -c 24/d'                  "${ptRoot}/parallelTestExecute.sh"
+    sed_ie '/#SBATCH -N 1/d'                   "${ptRoot}/parallelTestExecute.sh"
+    sed_ie '/#SBATCH -t 0-03:30/d'             "${ptRoot}/parallelTestExecute.sh"
+    sed_ie '/#SBATCH -p REQUESTED_PARTITION/d' "${ptRoot}/parallelTestExecute.sh"
+    sed_ie '/#SBATCH --mem=90000/d'            "${ptRoot}/parallelTestExecute.sh"
+    sed_ie '/#SBATCH --mail-type=END/d'        "${ptRoot}/parallelTestExecute.sh"
 
     # Replace "REQUESTED_PARTITION" with the partition name
-    sed_ie "${sedCmd}" "${ptRoot}/parTestCompile.sh"
-    sed_ie "${sedCmd}" "${ptRoot}/parTestExecute.sh"
+    sed_ie "${sedCmd}" "${ptRoot}/parallelTestCompile.sh"
+    sed_ie "${sedCmd}" "${ptRoot}/parallelTestExecute.sh"
 
     # Submit compilation tests script
-    output=$(bsub parTestCompile.sh)
+    output=$(bsub parallelTestCompile.sh)
     output=($output)
     cmpId=${output[1]}
     cmpId=${cmpId/<}
     cmpId=${cmpId/>}
 
     # Submit execution tests script as a job dependency
-    output=$(bsub -w "exit(${cmpId},0)" parTestExecute.sh)
+    output=$(bsub -w "exit(${cmpId},0)" parallelTestExecute.sh)
     output=($output)
     exeId=${output[1]}
     exeId=${exeId/<}
@@ -269,7 +263,7 @@ else
     # Run compilation tests
     echo ""
     echo "Compiliation tests are running..."
-    ./parTestCompile.sh &
+    ./parallelTestCompile.sh &
 
     # Change back to this directory
     cd "${thisDir}"